week_8/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
streamlit>=1.10.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.1
langchain-core>=0.1.0
faiss-cpu>=1.7.0
python-dotenv>=0.20.0
transformers>=4.20.0
torch>=1.10.0
accelerate>=0.10.0
datasets>=2.0.0
pdfminer.six>=20200517
pypdf>=3.0.0
chromadb>=0.3.0
tiktoken>=0.4.0
langchain-openai>=0.0.1
plotly>=5.10.0
google-generativeai>=0.3.0
langchain-google-genai>=0.0.1
huggingface-hub>=0.10.0 